requirements.txt
commonroad-io>=2018.1
opendrive2lanelet>=1.0.1